Lint: Arithmetic fixes

+++ app/src/main/java/com/tokenbrowser/crypto/cryptohash/DigestEngine.java
@@ -144,7 +144,7 @@ public abstract class DigestEngine implements Digest {
 
     public void update(byte input)
     {
-        inputBuf[inputLen ++] = (byte)input;
+        inputBuf[inputLen ++] = input;
         if (inputLen == blockLen) {
             processBlock(inputBuf);
             blockCount ++;
